Moved EditParcel tests
